Calculate isocubes coordinates from a height matrix

Calculate isocubes coordinates from a height matrix

Usage

coords_heightmap(
  mat,
  col = NULL,
  scale = 1,
  flipx = FALSE,
  flipy = TRUE,
  ground = "xz",
  solid = TRUE,
  verbose = FALSE
)

Arguments

mat

integer matrix. The matrix will be interpreted as cubes flat on the page, with the value in the matrix interpreted as the height above the page.

col

matrix of colours the same dimensions as the mat argument. Default: NULL. If col is not NULL, then a col column will be included in the final returned coordinates.

scale

scale factor for values in matrix. Default = 1

flipx, flipy

Should the matrix be flipped in the horizontal/vertical directions (respectively)? Default: flipx = FALSE, flipy = TRUE.

Note: flipy defaults to TRUE as matrices are indexed from the top-down, but the isometric coordinate space is increasing from the bottom up. Flipping the matrix vertically is usually what you want.

ground

Orientation of the ground plane. Default: 'xz'. Possible values 'xz', 'xy'

solid

Should the heightmap be made 'solid' i.e. without holes? default: TRUE. This can be an expensive operation in terms of both memory and CPU, but should be OK for simple examples. Set to FALSE if things take too long. This operation works by extruding cubes down from the top of the height map to the floor to ensure gaps do not appear when the slope is too great.

verbose

Be verbose? default: FALSE